pyproject.toml
[project]
name = "mcpagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.0.2",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.12",
    "langgraph>=1.0.2",
    "mcp>=1.20.0",
]
